Clean up warnings about the type mismatch
Clean up the type mismatch warnings to satisfy the compiler.

+++ single_include_test/csv.hpp
@@ -7774,7 +7774,7 @@ namespace csv {
                 return false;
             }
 
-            value += digit * static_cast<int>(pow(16, base16_exponent));
+            value += digit * static_cast<int>(pow(16, static_cast<int>(base16_exponent)));
             base16_exponent--;
         }
